+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v1.1.0 - 2022-05-25
+## feature: Add button for detail
+
+- Add button detail on every event
+- Add new route
+- Add new view: EventDetail
+- Update EventItem
+
+---
 # v1.0.0 - 2022-05-19
 ## feature: Init events with vue3